rename SqlNormalizer to SqlSanitizer (more descriptive).
