+++ javaagent-api/javaagent-api.gradle
@@ -12,14 +12,14 @@ project.ext.minimumInstructionCoverage = 0.0
 
 javacc {
   configs {
-    normalizer {
-      inputFile = file('src/main/javacc/SqlNormalizer.jj')
-      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.normalizer'
+    sanitizer {
+      inputFile = file('src/main/javacc/SqlSanitizer.jj')
+      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.sanitizer'
       outputDir = file("${project.buildDir}/generated/javacc/sql")
     }
     statmentInfoExtractor {
       inputFile = file('src/main/javacc/SqlStatementInfoExtractor.jj')
-      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.normalizer'
+      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.sanitizer'
       outputDir = file("${project.buildDir}/generated/javacc/sql")
     }
   }
