+++ javaagent-api/src/main/javacc/SqlStatementInfoExtractor.jj
@@ -11,7 +11,7 @@ options {
 
 PARSER_BEGIN(SqlStatementInfoExtractor)
 
-package io.opentelemetry.javaagent.instrumentation.api.db.normalizer;
+package io.opentelemetry.javaagent.instrumentation.api.db.sanitizer;
 
 import io.opentelemetry.javaagent.instrumentation.api.db.SqlStatementInfo;
 import java.io.StringReader;
